+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## Unreleased
 
+## 3.0.2 - 2021-04-29
+
+### Changed
+
+- Fixed broken pagination in PagerDuty integration. Pagination expected a
+  boolean response body property called `more`, but the property actually exists
+  in the response body at `data.more`.
+
 ## 3.0.1 - 2021-04-27
 
 ### Changed